pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "owhenrique",email = "me.pauloalmeida@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "fastapi[standard] (>=0.122.0,<0.123.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlmodel (>=0.0.27,<0.0.28)"
]

[tool.poetry]
packages = [{include = "app", from = "src"}]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_option]
pythonpath = ["src"]
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
dev = 'fastapi dev src/app/main.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=src/app -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "taskipy (>=1.14.1,<2.0.0)",
    "ruff (>=0.14.6,<0.15.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]
